pyproject.toml
[project]
name = "oqtopus-util"
version = "1.0.0"
description = "Template repository for building Python projects in the OQTOPUS ecosystem"
readme = "README.md"
license = "Apache-2.0"
authors = [{ name = "oqtopus-team", email = "oqtopus-team@googlegroups.com" }]
keywords = []
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "License :: OSI Approved :: Apache Software License",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering",
  "Topic :: Scientific/Engineering :: Physics",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Programming Language :: Python :: Implementation :: CPython",
]
requires-python = ">=3.11,<4.0"
dependencies = [
  "pyyaml>=6.0.3",
]

[project.urls]
documentation = "https://oqtopus-util.readthedocs.io/"
repository = "https://github.com/oqtopus-team/oqtopus-util"

[dependency-groups]
dev = [
  "mypy>=1.19.1",
  "pre-commit>=4.5.1",
  "ruff>=0.15.4",
  "types-pyyaml>=6.0.12.20250915",
]
test = [
  "pytest>=9.0.2",
  "pytest-cov>=7.0.0",
  "pytest-mock>=3.15.1",
]
docs = [
  "mkdocs<2.0", # Stay on 1.x to maintain plugin compatibility
  "mkdocs-gen-files>=0.6.0",
  "mkdocs-glightbox>=0.5.2",
  "mkdocs-jupyter>=0.25.1",
  "mkdocs-literate-nav>=0.6.2",
  "mkdocs-macros-plugin>=1.5.0",
  "mkdocs-material>=9.7.4",
  "mkdocstrings[python]>=1.0.3",
  "pymarkdownlnt>=0.9.35",
  "pymdown-extensions>=10.21",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/oqtopus_util"]

[tool.ruff]
preview = true
include = [
  "src/**/*.py",
  "tests/**/*.py",
]
exclude = [
]
lint.select = ["ALL"]
lint.ignore = [
  "COM812",
  "CPY001",
  "D100",
  "D104",
  "D107",
  "D203",
  "D213",
  "ISC001",
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
  "ANN201",
  "ANN205",
  "ANN401",
  "CPY001",
  "D",
  "PLC1901",
  "PLR2004",
  "PLR6301",
  "S101",
]

[tool.mypy]
files = ["src", "tests"]
exclude = [
]

[tool.pytest.ini_options]
addopts = [
  "--cov=src",
  "--cov-report=term-missing",
  "--cov-report=html",
  "--cov-report=xml",
]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.pymarkdown]
plugins.line-length.enabled = false # MD013: Disable line length limit
plugins.first-line-heading.enabled = false # MD041: Allow files to start without a top-level heading
